#126 refactor: 이미지컴포넌트 css 파일 분리
