+++ frontend/src/components/CheckImg.css
@@ -1,0 +1,24 @@
+.profileimgs {
+    position: relative;
+    display: inline-block;
+}
+
+.profileimg {
+    width: 150px;
+    height: 150px;
+    cursor: pointer;
+    transition: filter 0.3s ease;
+}
+
+.profileimg.darken {
+    filter: brightness(50%);
+}
+
+.checkmark {
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    background-color: white;
+    border-radius: 50%;
+    padding: 5px;
+}
