+++ frontend/src/components/LivingBtn.css
@@ -1,0 +1,23 @@
+.livingbtn {
+    width: 120px;
+    height: 120px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    border: 1px solid;
+    border-radius: 10px;
+    box-sizing: border-box;
+}
+
+.livingimg {
+    width: 70px;
+    height: 70px;
+    object-fit: cover;
+}
+
+.livingtext {
+    margin-top: 10px;
+    text-align: center;
+    font-size: 14px;
+}
