+++ frontend/src/components/LivingBtn.jsx
@@ -1,0 +1,12 @@
+import "./LivingBtn.css"
+
+const LivingBtn = ({ src, text }) => {
+  return (
+    <div className="livingbtn">
+      <img src={src} className="livingimg" />
+      <div className="livingtext">{text}</div>
+    </div>
+  );
+};
+
+export default LivingBtn;
